chore(providence-analytics): apply "custom projects" in find-exports
