+++ packages-node/providence-analytics/src/program/analyzers/find-exports.js
@@ -4,23 +4,24 @@ const { default: traverse } = require('@babel/traverse');
 const { Analyzer } = require('./helpers/Analyzer.js');
 const { trackDownIdentifier } = require('./helpers/track-down-identifier.js');
 const { normalizeSourcePaths } = require('./helpers/normalize-source-paths.js');
-const { aForEach } = require('../utils/async-array-utils.js');
 const { LogService } = require('../services/LogService.js');
 
 /**
  * @typedef {import('./helpers/track-down-identifier.js').RootFile} RootFile
  * @typedef {object} RootFileMapEntry
- * @property {string} currentFileSpecifier this is the local name in the file we track from
- * @property {RootFile} rootFile contains file(filePath) and specifier
+ * @typedef {string} currentFileSpecifier this is the local name in the file we track from
+ * @typedef {RootFile} rootFile contains file(filePath) and specifier
+ * @typedef {RootFileMapEntry[]} RootFileMap
+ *
+ * @typedef {{ exportSpecifiers:string[]; localMap: object; source:string, __tmp: { path:string } }} FindExportsSpecifierObj
  */
 
 /**
- * @typedef {RootFileMapEntry[]} RootFileMap
+ * @param {FindExportsSpecifierObj[]} transformedEntry
  */
-
 async function trackdownRoot(transformedEntry, relativePath, projectPath) {
   const fullCurrentFilePath = pathLib.resolve(projectPath, relativePath);
-  await aForEach(transformedEntry, async specObj => {
+  for (const specObj of transformedEntry) {
     /** @type {RootFileMap} */
     const rootFileMap = [];
     if (specObj.exportSpecifiers[0] === '[file]') {
@@ -39,7 +40,7 @@ async function trackdownRoot(transformedEntry, relativePath, projectPath) {
        *   }
        * }
        */
-      await aForEach(specObj.exportSpecifiers, async (/** @type {string} */ specifier) => {
+      for (const specifier of specObj.exportSpecifiers) {
         let rootFile;
         let localMapMatch;
         if (specObj.localMap) {
@@ -48,7 +49,7 @@ async function trackdownRoot(transformedEntry, relativePath, projectPath) {
         // TODO: find out if possible to use trackDownIdentifierFromScope
         if (specObj.source) {
           // TODO: see if still needed: && (localMapMatch || specifier === '[default]')
-          const importedIdentifier = (localMapMatch && localMapMatch.local) || specifier;
+          const importedIdentifier = localMapMatch?.local || specifier;
           rootFile = await trackDownIdentifier(
             specObj.source,
             importedIdentifier,
@@ -69,10 +70,10 @@ async function trackdownRoot(transformedEntry, relativePath, projectPath) {
           };
           rootFileMap.push(entry);
         }
-      });
+      }
     }
     specObj.rootFileMap = rootFileMap;
-  });
+  }
   return transformedEntry;
 }
 
@@ -139,6 +140,8 @@ function findExportsPerAstEntry(ast, { skipFileImports }) {
   LogService.debug(`Analyzer "find-exports": started findExportsPerAstEntry method`);
 
   // Visit AST...
+
+  /** @type {FindExportsSpecifierObj} */
   const transformedEntry = [];
   // Unfortunately, we cannot have async functions in babel traverse.
   // Therefore, we store a temp reference to path that we use later for
@@ -147,7 +150,7 @@ function findExportsPerAstEntry(ast, { skipFileImports }) {
     ExportNamedDeclaration(path) {
       const exportSpecifiers = getExportSpecifiers(path.node);
       const localMap = getLocalNameSpecifiers(path.node);
-      const source = path.node.source && path.node.source.value;
+      const source = path.node.source?.value;
       transformedEntry.push({ exportSpecifiers, localMap, source, __tmp: { path } });
     },
     ExportDefaultDeclaration(path) {
@@ -205,7 +208,8 @@ class FindExportsAnalyzer extends Analyzer {
      * Traverse
      */
     const projectPath = cfg.targetProjectPath;
-    const queryOutput = await this._traverse(async (ast, { relativePath }) => {
+
+    const traverseEntryFn = async (ast, { relativePath }) => {
       let transformedEntry = findExportsPerAstEntry(ast, cfg);
 
       transformedEntry = await normalizeSourcePaths(transformedEntry, relativePath, projectPath);
@@ -213,6 +217,12 @@ class FindExportsAnalyzer extends Analyzer {
       transformedEntry = cleanup(transformedEntry);
 
       return { result: transformedEntry };
+    };
+
+    const queryOutput = await this._traverse({
+      traverseEntryFn,
+      filePaths: cfg.targetFilePaths,
+      projectPath: cfg.targetProjectPath,
     });
 
     /**
